Accept DBI connection params, and support on-connect queries.
Applying patch from Igor Bujna ading the ability to provide parameters to be
passed to the DBI->connect call, e.g. RaiseError or AutoCommit, and also the
ability to provide a list of queries which should be performed immediately upon
connection to the database.

Thanks Igor!

In documenting these changes, I also improved the existing documentation
slightly.