Handle multiple named database connections.
This is thanks to Ignor Bujna who supplied the original code for this - thanks
Igor!

I've refactored it slightly and added some documentation.

This allows you to say database($name) where $name is the name of a connection
defined in your configuration file, so multiple connections are easily handled.

Since this is a fairly major change, I'll release it as a dev release for
testing first.