Dirty bodge to work around problems on Windows systems.
On some Windows systems, tests were failing due to the lack of views/public/lib
dirs within t/lib.

Credit to Sawyer for working out that this is due to Cwd on *nix systems using
an XS version, but Windows machines using a pure-Perl version which behaves
differently, croaking if a non-existent path is provided.

As Git can't track empty dirs, I've added a .placeholder file in each dir so
that I can actually check them in (bodgy goodness).