Add new auto UTF-8 support feature.
This resolves RT #65779 (also GH-6).

I'll release this as a developer release first and encourage testing.

Currently, I think it should be enabled by default if you're using charset
'UTF-8', as "Just Working" is the behaviour users would expect.  There is
however a config setting to disable it, should that be necessary for any reason.