Only log queries if log_queries is enabled in config.
Previously, queries generated by quick_select(), quick_insert() etc in
Dancer::Plugin::Database::Handle were logged (at debug level).

This could potentially lead to users unwittingly logging sensitive information
going to/from the database, so I've made it a configurable option, off by
default.

To enable logging of these queries, set the log_queries option to a true value.

This would provide a very minor performance boost too, as without the option
enabled, Dancer::Logger::debug() won't be called at all.

The setting is read at connection time and stored with the database handle, so
changing this setting during runtime but after connection to the database will
have no effect on whether queries are logged; the value as it was when the
connection was obtained will continue to be used.