Extract driver from DSN so we can still do auto utf8.
If we're given a pre-assembled DSN instead of a set of driver/host/database etc
params, extract the driver from the DSN, so we don't trigger undefined var
warnings, and more importantly, so the auto UTF-8 support can be enabled if the
app has charset: UTF-8 set.

Thanks for Matthew Vickers (@mvickers) for bringing this to my attention!