Handle "is" operator in where clauses properly - Issue 15.
Installing Dancer::Plugin::Database on a VPS, I got a test failure, detailed in
Issue #15.

Refactored _get_where_sql() to return a value indicating whether or not a bind
param is being used, and to simply return "IS NULL" / "IS NOT NULL" depending on
the negation parameter.  Also, warns if a non-undef value was passed, as that
doesn't make sense - you'll always want to test for nullness by providing undef
as the value, e.g.:

    foo => { is => undef },
    bar => { is => undef, not => 1 },

It doesn't make sense to pass a non-undef value, e.g.:

    foo => { is => 42 }

(That should be 'eq', or, more succintly, `foo => 42` (although the longer form
would need to be used to negate it, e.g. `foo => { ne => 42 }`)