Port to Dancer 2
*** Note: you should upgrade Dancer2 to the latest version on GitHub first. ***

Main changes are the use of the $dsl object instead of calling Dancer's
keywords directly in the plugin's code.

The test has been reorganized a bit as well, because the config is now scoped
in Dancer 2 and settings defined in a test script are not visible in
t::lib::TestApp (which is normal).