Cache a ref to settings hashref so it doesn't go out of scope.
This should fix Debian RT #665221:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=665221

The problem showed only in certain circumstances:

(a) a hashref of settings was passed to database(), rather than the name of a
connection from the config file (or nothing)
(b) that hashref then goes out of scope, and another hashref of DB connection
settings happens to re-use the address the first had, leading us to incorrectly
return the handle we obtained before.

I don't suppose this has ever showed up in real-world usage, but was caught by
the test suite, which, in the course of testing the ability to pass settings as
hashrefs, does exactly this (but, annoyingly, it only sometimes happens).

Thanks to the Debian Perl team for reporting this to me and prodding me to fix
it, and thanks to Sam Kington (@skington) for the cunning idea of simply storing
a ref to the original settings hashref to avoid it going out of scope.

Note that this may not be a 100% perfect test; I think you could still get
undesired results if you purposely change the settings within the hashref after
first using it to get a connection, but I think defending against that would be
adding inefficiency and complexity to handle a situation which is incredibly
unlikely in real-world usage.  Also, if you really, really, really want to shoot
yourself in the foot, who am I to tell you you can't?  It's your own foot, after
all.