Allow a false value for connection_check_threshold to disable checking.
At the moment, if connection_check_threshold is not supplied, or has a false
value, we default to 30 seconds.  We should default it it isn't supplied, but
if it's supplied with a false value, we should take that to mean that we should
not perform connectivity checks.

This addresses the point raised in GH #22 - that connectivity checking cannot be
disabled - which may be desireable if using the DBI `mysql_auto_reconnect`
attribute instead.